requirements.txt
fastapi~=0.88.0
uvicorn~=0.20.0
validators~=0.20.0
pydantic~=1.10.4
APScheduler~=3.9.1.post1
requests~=2.28.1
loguru~=0.6.0
pytest~=7.2.0
psycopg2-binary~=2.9.5